Update _.every and adjust for single value arrays
Code described in the previous commit and in the pseudocode has been
added. Additionally, the current implementation of the function was
having trouble handling arrays of length 1. In order to reconcile this,
reduce is given a default accumulator of true. With this, it can
evaluate every element of the array in order to determine if it is true
or false.

  1. Previously the function conducted a truthy test on an undefined
  accumulator and a single value – which always resolved to false
  2. Now the accumulator always begins as true, which is compared along
  with the first element of the array in a truthy test