Add code for _.shuffle according to pseudocode
Additional clarification on logic:
By putting the indexes of the original array into its own array, we
can randomly generate a value from this array using Math.random() *
the indexes array's length. We can similarly do this for the values
array. By selecting a random index and a random value, we can ensure
that the values we are inserting into the shuffledArr are indeed random.
We also ensure that there are no duplicates in the shuffledArr by
splicing those values from the respective indexes and values array after
inserting them into the shuffledArr